General tidy up.
