+++ Sources/SwiftUICharts/LineChart/Models/PointStyle.swift
