+++ Sources/SwiftUICharts/Shared/Extras/Extensions.swift
@@ -22,7 +22,6 @@ extension View {
                                                          if ifTransform: (Self) -> TrueContent,
                                                          else elseTransform: (Self) -> FalseContent
     ) -> some View {
-        
         if condition {
             ifTransform(self)
         } else {
