+++ Sources/SwiftUICharts/SharedLineAndBar/Models/ChartViewData.swift
