+++ Sources/SwiftUICharts/SharedLineAndBar/Shapes/LabelShape.swift
