+++ Sources/SwiftUICharts/SharedLineAndBar/ViewModifiers/XAxisGrid.swift
@@ -58,40 +58,3 @@ extension View {
         self.modifier(XAxisGrid(chartData: chartData))
     }
 }
-
-
-internal struct VerticalGridView<T>: View where T: LineAndBarChartData {
-    
-    var chartData : T
-    
-    @State var startAnimation : Bool = false
-    
-    var body: some View {
-        VerticalGridShape()
-            .trim(to: startAnimation ? 1 : 0)
-            .stroke(chartData.chartStyle.xAxisGridStyle.lineColour,
-                    style: StrokeStyle(lineWidth: chartData.chartStyle.xAxisGridStyle.lineWidth,
-                                       dash     : chartData.chartStyle.xAxisGridStyle.dash,
-                                       dashPhase: chartData.chartStyle.xAxisGridStyle.dashPhase))
-            .frame(width: chartData.chartStyle.xAxisGridStyle.lineWidth)
-            .animateOnAppear(using: chartData.chartStyle.globalAnimation) {
-                self.startAnimation = true
-            }
-            .animateOnDisappear(using: chartData.chartStyle.globalAnimation) {
-                self.startAnimation = false
-            }
-    }
-}
-internal struct VerticalGridShape: Shape {
-    
-    internal func path(in rect: CGRect) -> Path {
-        
-        var path  = Path()
-
-        path.move(to: CGPoint(x: 0, y: rect.height))
-        path.addLine(to: CGPoint(x: 0, y: 0))
-        
-        return path
-    }
-    
-}
