+++ Sources/SwiftUICharts/SharedLineAndBar/ViewModifiers/YAxisGrid.swift
@@ -17,9 +17,7 @@ internal struct YAxisGrid<T>: ViewModifier where T: LineAndBarChartData {
                 VStack {
                     ForEach((0...chartData.chartStyle.yAxisGridStyle.numberOfLines), id: \.self) { index in
                         if index != 0 {
-                            
                             HorizontalGridView(chartData: chartData)
-                            
                             Spacer()
                                 .frame(minHeight: 0, maxHeight: 500)
                         }
@@ -58,41 +56,5 @@ extension View {
     */
     public func yAxisGrid<T: LineAndBarChartData>(chartData: T) -> some View {
         self.modifier(YAxisGrid<T>(chartData: chartData))
-    }
-}
-
-
-internal struct HorizontalGridView<T>: View where T: LineAndBarChartData {
-    
-    var chartData : T
-    
-    @State var startAnimation : Bool = false
-    
-    var body: some View {
-        HorizontalGridShape()
-            .trim(to: startAnimation ? 1 : 0)
-            .stroke(chartData.chartStyle.yAxisGridStyle.lineColour,
-                    style: StrokeStyle(lineWidth: chartData.chartStyle.yAxisGridStyle.lineWidth,
-                                       dash     : chartData.chartStyle.yAxisGridStyle.dash,
-                                       dashPhase: chartData.chartStyle.yAxisGridStyle.dashPhase))
-            .frame(height: chartData.chartStyle.yAxisGridStyle.lineWidth)
-            .animateOnAppear(using: chartData.chartStyle.globalAnimation) {
-                self.startAnimation = true
-            }
-            .animateOnDisappear(using: chartData.chartStyle.globalAnimation) {
-                self.startAnimation = false
-            }
-    }
-}
-internal struct HorizontalGridShape: Shape {
-    
-    internal func path(in rect: CGRect) -> Path {
-        
-        var path  = Path()
-
-        path.move(to: CGPoint(x: 0, y: 0))
-        path.addLine(to: CGPoint(x: rect.width, y: 0))
-        
-        return path
     }
 }
