+++ Sources/SwiftUICharts/SharedLineAndBar/Views/HorizontalGridView.swift
@@ -1,0 +1,31 @@
+//
+//  HorizontalGridView.swift
+//  
+//
+//  Created by Will Dale on 08/02/2021.
+//
+
+import SwiftUI
+
+internal struct HorizontalGridView<T>: View where T: LineAndBarChartData {
+    
+    var chartData : T
+    
+    @State var startAnimation : Bool = false
+    
+    var body: some View {
+        HorizontalGridShape()
+            .trim(to: startAnimation ? 1 : 0)
+            .stroke(chartData.chartStyle.yAxisGridStyle.lineColour,
+                    style: StrokeStyle(lineWidth: chartData.chartStyle.yAxisGridStyle.lineWidth,
+                                       dash     : chartData.chartStyle.yAxisGridStyle.dash,
+                                       dashPhase: chartData.chartStyle.yAxisGridStyle.dashPhase))
+            .frame(height: chartData.chartStyle.yAxisGridStyle.lineWidth)
+            .animateOnAppear(using: chartData.chartStyle.globalAnimation) {
+                self.startAnimation = true
+            }
+            .animateOnDisappear(using: chartData.chartStyle.globalAnimation) {
+                self.startAnimation = false
+            }
+    }
+}
